refactor: move aggregate statistics to datafusion-physical-optimizer
