+++ datafusion/physical-optimizer/Cargo.toml
@@ -34,5 +34,6 @@ workspace = true
 [dependencies]
 datafusion-common = { workspace = true, default-features = true }
 datafusion-execution = { workspace = true }
+datafusion-expr = { workspace = true }
 datafusion-physical-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
