+++ datafusion/physical-optimizer/src/lib.rs
@@ -17,6 +17,7 @@
 // Make cheap clones clear: https://github.com/apache/datafusion/issues/11143
 #![deny(clippy::clone_on_ref_ptr)]
 
+pub mod aggregate_statistics;
 mod optimizer;
 pub mod output_requirements;
